Fix ORDER BY for all_primes.html
git-svn-id: https://svn.code.sf.net/p/prpnet/code@131 029e2237-403a-42e8-9450-7768c132b01b

+++ source/upgrade_to_5.6.0.sql
@@ -1,7 +1,7 @@
 alter table UserPrimes add k bigint;
-alter table UserPrimes add b bigint;
-alter table UserPrimes add n bigint;
-alter table UserPrimes add c bigint;
+alter table UserPrimes add b int;
+alter table UserPrimes add n int;
+alter table UserPrimes add c int;
 
 update UserPrimes up set k = (select c.k from Candidate c where c.candidateName = up.candidateName);
 update UserPrimes up set b = (select c.b from Candidate c where c.candidateName = up.candidateName);
